Daemon stability improvements

+++ Source/Azure iOS UI/Definitions/SearchObject.h
@@ -25,7 +25,7 @@
 - (BOOL)isDecimalNumberSearch;
 - (BOOL)isByteSearch;
 
-- (int)getSearchSize;
+- (NSInteger)getSearchSize;
 - (void*)getRawData;
 
 - (NSString *)toString;
